+++ src/components/forms/text-area/text-area.stories.mdx
@@ -1,30 +1,24 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import { TextInput } from './text-input';
-import { getSampleController } from './text-field-story-helper';
-import { SIMPLE_EMAIL_PATTERN } from '../constants';
+import { TextArea } from './text-area';
+import { getSampleController } from '../text-field-story-helper';
 
 <Meta
-  title="Components/Forms/TextInput"
-  component={{ TextInput }} />
+  title="Components/Forms/TextArea"
+  component={{ TextArea }} />
 
-# TextInput
+# TextArea
 The component is used to get input from the user. Besides the `label` it receives props of the type `UseControllerProps`. The API of this component is based on the [controlled approach of react-hook-form](https://react-hook-form.com/api/usecontroller).
 
 <Canvas withSource='open'>
   <Story name='Regular'>
-    <TextInput
-      label="E-Mail"
+    <TextArea
+      label="Message"
       name="message"
       control={getSampleController()}
-      rules={{
-        required: "Your e-mail is missing", pattern: {
-          value: SIMPLE_EMAIL_PATTERN,
-          message: "Something is wrong with your e-mail",
-        },
-      }}
+      rules={{ required: "Your message is missing" }}
       style={{width: 'calc(100% - 30px)'}}
     />
   </Story>
 </Canvas>
 
-<ArgsTable of={TextInput} />
+<ArgsTable of={TextArea} />
