+++ src/components/forms/text-input/text-input.stories.mdx
@@ -1,24 +1,30 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import { TextArea } from './text-area';
-import { getSampleController } from './text-field-story-helper';
+import { TextInput } from './text-input';
+import { getSampleController } from '../text-field-story-helper';
+import { SIMPLE_EMAIL_PATTERN } from '../constants';
 
 <Meta
-  title="Components/Forms/TextArea"
-  component={{ TextArea }} />
+  title="Components/Forms/TextInput"
+  component={{ TextInput }} />
 
-# TextArea
+# TextInput
 The component is used to get input from the user. Besides the `label` it receives props of the type `UseControllerProps`. The API of this component is based on the [controlled approach of react-hook-form](https://react-hook-form.com/api/usecontroller).
 
 <Canvas withSource='open'>
   <Story name='Regular'>
-    <TextArea
-      label="Message"
+    <TextInput
+      label="E-Mail"
       name="message"
       control={getSampleController()}
-      rules={{ required: "Your message is missing" }}
+      rules={{
+        required: "Your e-mail is missing", pattern: {
+          value: SIMPLE_EMAIL_PATTERN,
+          message: "Something is wrong with your e-mail",
+        },
+      }}
       style={{width: 'calc(100% - 30px)'}}
     />
   </Story>
 </Canvas>
 
-<ArgsTable of={TextArea} />
+<ArgsTable of={TextInput} />
