+++ src/components/pages/contact/contact.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import { Address } from './address';
-import { Form } from './form';
+import { ContactForm } from './contact-form';
 
 <Meta title="Components/Pages/Contact" component={{ Image }} />
 
@@ -17,6 +17,6 @@ The contact page consists of two sections that allow the user to write a message
 ## Form
 <Canvas withSource={'open'}>
   <Story name="Form">
-    <Form/>
+    <ContactForm/>
   </Story>
 </Canvas>
